Update README.md
Total points: 14.

For this assignment, we will again be using the hr database from HW9.

The following is the structure of the hr database:

*Provided by w3resource website

It’s good practice to make SQL keywords uppercase to distinguish them from other parts of your query. Please do this to make it easier for the grader. Also remember to include a semicolon at the end of your query. This tells SQL where the end of your query is!

    [1.5 pts] Provide the SQL code to display the last day of the month (in date format) three months before the current month.

    [1 pt] Provide the SQL code to display the current date in the given format: e.g. 11/29/2021

    [1.5 pts] Provide the SQL code to display the job ID and maximum salary of the employees where maximum salary is greater than or equal to $4000.